requirements.txt
flask
flask-cors
pillow
torch
torchvision
albumentations
python-dotenv
opencv-python
numpy
werkzeug